update the snapshot, not really sure what changed though
To clarify, I'm not sure what changed in the code to cause the changed snapshot.
I tried git bisecting back to when I added snapshots in the first place, but all
of the commits were bad. I'm already thinking these might be a bad idea.

the actual difference is in the filters at the end. I think the new one only had
the original filters, and before I just kept accumulating filters on top.
something like that. One of them had 10 filters and the other 6